update: move component and component list to models
